Fixes crashes when we modify dependencies in ewpi packages
The crash happens because the _ew_package_index is not filled for
dependencies not found in the current compilation, but array is
iterated over all indices anyway, instead of just the dependencies dst
count.

The array returns garbage which is then dereferenced in _ewpi_pkgs and
crashes.

This patch makes this iterate over just the dependencies instead, and as
such doesn't iterate over uninitialized array elements.